🚚 Moved and renamed `Box` component
