+++ src/components/layouts/partials/home-header.js
@@ -3,7 +3,6 @@ import NavBar from './navbar'
 // import SectionHeading from './section-heading'
 import SearchModal from '../../global/search-modal'
 import Icon from '../../global/icon'
-// import Box from '../../layouts/partials/box'
 import { Spirit } from '../../spirit-styles'
 import PropTypes from 'prop-types'
 import { Link } from 'gatsby'
