+++ src/pages/index.js
@@ -6,7 +6,7 @@ import PropTypes from 'prop-types'
 import Icon from '../components/global/icon'
 import Layout from '../components/layouts/default'
 import { Spirit } from '../components/spirit-styles'
-import Box from '../components/layouts/partials/box'
+import Box from '../components/Box'
 import { MetaData } from '../components/meta'
 import HomeHeader from '../components/layouts/partials/home-header'
 import getMetaImageUrls from '../utils/getMetaImageUrls'
